moved SPDHG to file
